Refactor and renamed LevelViewTwo to LevelViewBoss
